fix: report feature resolution errors
Some packages may have bugs in their feature configuration. One case is
async-std - async-std 1.12.0 has an optional dependency on surf 2.0.0.
However surf 2.0.0 has a required dependency on async-std 1.6.0. When we
enable the dep:surf feature, we encounter a circular dependency. Note
that this issue is also present if you perform a `cargo add async-std -F
surf` in an empty test project.

The suggested solution online appears to be to manually identify the
circular dependency, and fix package versions to resolve the issue. In
the above async-std case, this would be pinning async-std to 1.6.0.
However, this solution isn't suitable for this use case due to its
manual nature.